Core: * Reorganize packages (move from "artifact.intrinsic" to "artifact.common" package). + Introduce PathBuilder helper class for work with absolute paths inside intrinsic objects. Intrinsics: + Add "Adapters" built-in object to hold convenience adapters identifiers. + Add "Adapters: by string" convenience adapter identifier for string-to-value parsers. + Add "Number@@ by string" adapter for string-to-number conversion. + Implement "Float@@ by string" adapter for string-to-float conversion. + Implement "Integer@@ by string" adapter for string-to-integer conversion.

+++ org.o42a.rt/include/o42a/integer.h
@@ -28,7 +28,7 @@ extern "C" {
 #endif
 
 
-void o42a_int_by_str(o42a_val_t *, const o42a_val_t *, uint8_t);
+void o42a_int_by_str(o42a_val_t *, const o42a_val_t *, uint32_t);
 
 
 #ifdef __cplusplus
